requirements.txt
feedparser==6.0.11
Jinja2==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.1
